#23 new method init with typepath in WP8
